fix error with resnext init and fix densenet18 init

+++ deepards/models/senet.py
@@ -153,7 +153,7 @@ class SEResNeXtBottleneck(Bottleneck):
     def __init__(self, inplanes, planes, groups, reduction, stride=1,
                  downsample=None, base_width=4):
         super(SEResNeXtBottleneck, self).__init__()
-        width = math.floor(planes * (base_width / 64)) * groups
+        width = int(math.floor(planes * (base_width / 64)) * groups)
         self.conv1 = nn.Conv1d(inplanes, width, kernel_size=1, bias=False,
                                stride=1)
         self.bn1 = nn.BatchNorm1d(width)
